Test terminating running jobs
